New job-template, job-group for horizon to test the rebase fix
job-template - "{name}-jiocloud-rebase-fix-testing"
* A rebase would have rewritten the branch history so...
removed the code to that does 'git pull' and...
if 'jiocloud_master' branch existed, delete it and fetch the latest and...
do a fresh checkout for every run of the job.
* Push the rebased code irrespective of whether tox tests pass.
This is temporary, until we figure out a way to resolve
the tox tests issue with our jio-master

Other than the above change, the rest is just a copy of the existing template
"{name}-jiocloud".